automation_inspector/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
PyYAML==6.0.1
